pyproject.toml
[project]
name = "exocet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nidaqmx>=1.2.0",
    "pyserial>=3.5",
    "questdb[dataframe]>=3.0.0",
    "streamlit>=1.50.0",
]
